Make some changes again that makes the game more interactive
Here in the updated version of the code we can see that we are counting the scores and displaying them after each round using alert. And at the end of the game we announce who has won the game. Here in the rounds that same selection is drawn we don't give and points and move onto the other rounds as usual.